feat: Separate the Card component from AllCourses.
