+++ client/src/pages/AllCourses/index.jsx
@@ -1,20 +1,12 @@
 import React, { useEffect, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import CourseService from 'services/course.service';
-import { ROUTER_PATH } from 'App';
-
 // Font Awesome
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faMagnifyingGlass } from '@fortawesome/free-solid-svg-icons';
-
-import Button from 'components/Button';
+import Card from 'components/Card';
 import './allCourses.scss';
 
 const AllCourses = ({ currentUser, setIsModalOpen }) => {
-  const navigate = useNavigate();
-  const handleTakeToLogin = () => {
-    setIsModalOpen(true);
-  };
   // state 用來儲存從API當中所獲得的Course Data
   const [courseData, setCourseData] = useState([]);
   const [searchInput, setSearchInput] = useState('');
@@ -50,18 +42,6 @@ const AllCourses = ({ currentUser, setIsModalOpen }) => {
     setSearchInput(''); // 清空搜尋欄位的值
   };
 
-  const handleTakeToEnroll = (course) => {
-    if (currentUser?.user.role === 'student') {
-      navigate(ROUTER_PATH.courseContent, {
-        state: course,
-      });
-    } else if (currentUser?.user.role === 'instructor') {
-      window.alert('Only students can enroll in courses. Please use student account to login.');
-    } else {
-      handleTakeToLogin();
-    }
-  };
-
   // 網頁組件渲染完後就執行effect
   useEffect(() => {
     // console.log('Using effect!');
@@ -93,56 +73,13 @@ const AllCourses = ({ currentUser, setIsModalOpen }) => {
             </div>
           )}
           {showMessage && <div className='message'>Please enter the course name you want to search.</div>}
-          {searchResult &&
-            searchResult.map((course) => (
-              <div className='course' key={course._id}>
-                <div className='intro'>
-                  <img src={course.img.replace('.png', 'l.png')} alt={`${course.title}-img`} />
-                  <h2>{course.title}</h2>
-                  <p>{course.subtitle}</p>
-                </div>
-                <div className='instructor'>{course.instructor.username}</div>
-                <div className='course-price'>
-                  <span>$ </span>
-                  <span>{course.price}</span>
-                </div>
-                <Button
-                  cx='see-more-btn'
-                  onClick={() => {
-                    handleTakeToEnroll(course);
-                  }}
-                >
-                  See More...
-                </Button>
-              </div>
-            ))}
+          {searchResult && searchResult.map((course) => <Card course={course} currentUser={currentUser} setIsModalOpen={setIsModalOpen}></Card>)}
         </div>
         {searchResult && searchResult.length > 0 && <div className='dividing-line'></div>}
 
         <div className='all-courses'>
           {courseData.map((course) => {
-            return (
-              <div className='course' key={course._id}>
-                <div className='intro'>
-                  <img src={course.img.replace('.png', 'l.png')} alt={`${course.title}-img`} />
-                  <h2>{course.title}</h2>
-                  <p>{course.subtitle}</p>
-                </div>
-                <div className='instructor'>{course.instructor.username}</div>
-                <div className='course-price'>
-                  <span>$ </span>
-                  <span>{course.price}</span>
-                </div>
-                <Button
-                  cx='see-more-btn'
-                  onClick={() => {
-                    handleTakeToEnroll(course);
-                  }}
-                >
-                  See More...
-                </Button>
-              </div>
-            );
+            return <Card course={course} currentUser={currentUser} setIsModalOpen={setIsModalOpen}></Card>;
           })}
         </div>
       </section>
